feat: remove walletService from accountsService
