+++ src/background/services/wallet/WalletService.ts
@@ -22,12 +22,7 @@ import {
   createWalletPolicy,
   DerivationPath,
   getAddressDerivationPath,
-  getAddressFromXPub,
   getAddressPublicKeyFromXPub,
-  getBech32AddressFromXPub,
-  getBtcAddressFromPubKey,
-  getEvmAddressFromPubKey,
-  getPubKeyFromTransport,
   getPublicKeyFromPrivateKey,
   getWalletFromMnemonic,
   JsonRpcBatchInternal,
@@ -43,9 +38,8 @@ import {
 } from '@metamask/eth-sig-util';
 import { LedgerService } from '../ledger/LedgerService';
 import { BaseWallet, Wallet } from 'ethers';
-import { networks, Transaction } from 'bitcoinjs-lib';
+import { Transaction } from 'bitcoinjs-lib';
 import { prepareBtcTxForLedger } from './utils/prepareBtcTxForLedger';
-import { ImportData, ImportType } from '../accounts/models';
 import ensureMessageIsValid from './utils/ensureMessageFormatIsValid';
 import { KeystoneWallet } from '../keystone/KeystoneWallet';
 import { KeystoneService } from '../keystone/KeystoneService';
@@ -797,173 +791,6 @@ export class WalletService implements OnLock, OnUnlock {
     this.eventEmitter.on(event, callback);
   }
 
-  async addAddress(
-    index: number,
-    walletId: string
-  ): Promise<Record<NetworkVMType, string>> {
-    const secrets = await this.secretService.getWalletAccountsSecretsById(
-      walletId
-    );
-
-    if (
-      secrets.secretType === SecretType.LedgerLive &&
-      !secrets.pubKeys[index]
-    ) {
-      // With LedgerLive, we don't have xPub or Mnemonic, so we need
-      // to get the new address pubkey from the Ledger device.
-      if (!this.ledgerService.recentTransport) {
-        throw new Error('Ledger transport not available');
-      }
-
-      // Get EVM public key from transport
-      const addressPublicKeyC = await getPubKeyFromTransport(
-        this.ledgerService.recentTransport,
-        index,
-        secrets.derivationPath
-      );
-
-      // Get X/P public key from transport
-      const addressPublicKeyXP = await getPubKeyFromTransport(
-        this.ledgerService.recentTransport,
-        index,
-        secrets.derivationPath,
-        'AVM'
-      );
-
-      if (
-        !addressPublicKeyC ||
-        !addressPublicKeyC.byteLength ||
-        !addressPublicKeyXP ||
-        !addressPublicKeyXP.byteLength
-      ) {
-        throw new Error('Failed to get public key from device.');
-      }
-
-      const pubKeys = [...(secrets?.pubKeys || [])];
-      pubKeys[index] = {
-        evm: addressPublicKeyC.toString('hex'),
-        xp: addressPublicKeyXP.toString('hex'),
-      };
-
-      await this.secretService.updateSecrets(
-        {
-          pubKeys,
-        },
-        walletId
-      );
-    }
-
-    if (secrets.secretType === SecretType.Seedless && !secrets.pubKeys[index]) {
-      const wallet = new SeedlessWallet({
-        networkService: this.networkService,
-        sessionStorage: new SeedlessTokenStorage(this.secretService),
-        addressPublicKey: secrets.pubKeys[0],
-      });
-
-      // Prompt Core Seedless API to derive new keys
-      await wallet.addAccount(index);
-      // Update the public keys in wallet
-      await this.secretService.updateSecrets(
-        {
-          pubKeys: await wallet.getPublicKeys(),
-        },
-        walletId
-      );
-    }
-
-    return this.getAddresses(index, walletId);
-  }
-
-  async getAddresses(
-    index: number,
-    walletId: string
-  ): Promise<Record<NetworkVMType, string> | never> {
-    if (!walletId) {
-      throw new Error('Wallet id not provided');
-    }
-
-    const secrets = await this.secretService.getWalletAccountsSecretsById(
-      walletId
-    );
-
-    if (!secrets) {
-      throw new Error('Wallet is not initialized');
-    }
-
-    const isMainnet = this.networkService.isMainnet();
-    const provXP = await this.networkService.getAvalanceProviderXP();
-
-    if (
-      secrets.secretType === SecretType.Ledger ||
-      secrets.secretType === SecretType.Mnemonic ||
-      secrets.secretType === SecretType.Keystone
-    ) {
-      // C-avax... this address uses the same public key as EVM
-      const cPubkey = getAddressPublicKeyFromXPub(secrets.xpub, index);
-      const cAddr = provXP.getAddress(cPubkey, 'C');
-
-      let xAddr, pAddr;
-      // We can only get X/P addresses if xpubXP is set
-      if (secrets.xpubXP) {
-        // X and P addresses different derivation path m/44'/9000'/0'...
-        const xpPub = Avalanche.getAddressPublicKeyFromXpub(
-          secrets.xpubXP,
-          index
-        );
-        xAddr = provXP.getAddress(xpPub, 'X');
-        pAddr = provXP.getAddress(xpPub, 'P');
-      }
-
-      return {
-        [NetworkVMType.EVM]: getAddressFromXPub(secrets.xpub, index),
-        [NetworkVMType.BITCOIN]: getBech32AddressFromXPub(
-          secrets.xpub,
-          index,
-          isMainnet ? networks.bitcoin : networks.testnet
-        ),
-        [NetworkVMType.AVM]: xAddr,
-        [NetworkVMType.PVM]: pAddr,
-        [NetworkVMType.CoreEth]: cAddr,
-      };
-    }
-
-    if (
-      secrets.secretType === SecretType.LedgerLive ||
-      secrets.secretType === SecretType.Seedless
-    ) {
-      // pubkeys are used for LedgerLive derivation paths m/44'/60'/n'/0/0
-      // and for X/P derivation paths  m/44'/9000'/n'/0/0
-      const addressPublicKey = secrets.pubKeys[index];
-
-      if (!addressPublicKey?.evm) {
-        throw new Error('Account not added');
-      }
-
-      const pubKeyBuffer = Buffer.from(addressPublicKey.evm, 'hex');
-
-      // X/P addresses use a different public key because derivation path is different
-      let addrX, addrP;
-      if (addressPublicKey.xp) {
-        const pubKeyBufferXP = Buffer.from(addressPublicKey.xp, 'hex');
-        addrX = provXP.getAddress(pubKeyBufferXP, 'X');
-        addrP = provXP.getAddress(pubKeyBufferXP, 'P');
-      }
-
-      return {
-        [NetworkVMType.EVM]: getEvmAddressFromPubKey(pubKeyBuffer),
-        [NetworkVMType.BITCOIN]: getBtcAddressFromPubKey(
-          pubKeyBuffer,
-          isMainnet ? networks.bitcoin : networks.testnet
-        ),
-        [NetworkVMType.AVM]: addrX,
-        [NetworkVMType.PVM]: addrP,
-        [NetworkVMType.CoreEth]: provXP.getAddress(pubKeyBuffer, 'C'),
-      };
-    }
-
-    throw new Error('No public key available');
-  }
-
   async getAddressesByIndices(
     indices: number[],
     chainAlias: 'X' | 'P',
@@ -991,137 +818,6 @@ export class WalletService implements OnLock, OnUnlock {
     );
   }
 
-  async addImportedWallet(importData: ImportData) {
-    const id = crypto.randomUUID();
-
-    // let the AccountService validate the account's uniqueness and save the secret using this callback
-    const commit = async () => {
-      // Need to re-map `data` to `secret` for private key imports
-      switch (importData.importType) {
-        case ImportType.PRIVATE_KEY:
-          await this.secretService.saveImportedWallet(id, {
-            secretType: SecretType.PrivateKey,
-            secret: importData.data,
-          });
-          break;
-        case ImportType.FIREBLOCKS:
-          await this.secretService.saveImportedWallet(id, {
-            secretType: SecretType.Fireblocks,
-            ...importData.data,
-          });
-          break;
-        case ImportType.WALLET_CONNECT:
-          await this.secretService.saveImportedWallet(id, {
-            secretType: SecretType.WalletConnect,
-            ...importData.data,
-          });
-          break;
-      }
-    };
-
-    if (
-      importData.importType === ImportType.FIREBLOCKS ||
-      importData.importType === ImportType.WALLET_CONNECT
-    ) {
-      return {
-        account: {
-          id,
-          ...importData.data.addresses,
-        },
-        commit,
-      };
-    }
-
-    if (importData.importType === ImportType.PRIVATE_KEY) {
-      return {
-        account: {
-          id,
-          ...this.#calculateAddressesForPrivateKey(importData.data),
-        },
-        commit,
-      };
-    }
-
-    throw new Error('Unknown import type');
-  }
-
-  async getImportedAddresses(id: string) {
-    const secrets = await this.secretService.getImportedAccountSecrets(id);
-
-    if (
-      secrets.secretType === SecretType.WalletConnect ||
-      secrets.secretType === SecretType.Fireblocks
-    ) {
-      return secrets.addresses;
-    }
-
-    if (secrets.secretType === SecretType.PrivateKey) {
-      return this.#calculateAddressesForPrivateKey(secrets.secret);
-    }
-
-    throw new Error('Unsupported import type');
-  }
-
-  #calculateAddressesForPrivateKey(privateKey: string) {
-    const addresses = {
-      addressBTC: '',
-      addressC: '',
-      addressAVM: '',
-      addressPVM: '',
-      addressCoreEth: '',
-    };
-
-    const isMainnet = this.networkService.isMainnet();
-    const provXP = isMainnet
-      ? Avalanche.JsonRpcProvider.getDefaultMainnetProvider()
-      : Avalanche.JsonRpcProvider.getDefaultFujiProvider();
-
-    try {
-      const publicKey = getPublicKeyFromPrivateKey(privateKey);
-      addresses.addressC = getEvmAddressFromPubKey(publicKey);
-      addresses.addressBTC = getBtcAddressFromPubKey(
-        publicKey,
-        isMainnet ? networks.bitcoin : networks.testnet
-      );
-      addresses.addressAVM = provXP.getAddress(publicKey, 'X');
-      addresses.addressPVM = provXP.getAddress(publicKey, 'P');
-      addresses.addressCoreEth = provXP.getAddress(publicKey, 'C');
-    } catch (err) {
-      throw new Error('Error while calculating addresses');
-    }
-
-    if (
-      !addresses.addressC ||
-      !addresses.addressBTC ||
-      !addresses.addressAVM ||
-      !addresses.addressPVM ||
-      !addresses.addressCoreEth
-    ) {
-      throw new Error(`Missing address`);
-    }
-
-    return addresses;
-  }
-
-  async deleteImportedWallets(ids: string[]) {
-    const deleted = await this.secretService.deleteImportedWallets(ids);
-
-    Object.values(deleted).forEach(async (wallet) => {
-      if (
-        wallet?.secretType === SecretType.WalletConnect ||
-        wallet?.secretType === SecretType.Fireblocks
-      ) {
-        await this.walletConnectService.deleteSession(
-          wallet.addresses.addressC
-        );
-      }
-    });
-  }
-
-  async deletePrimaryWallets(ids: string[]) {
-    await this.secretService.deletePrimaryWallets(ids);
-  }
-
   private async parseWalletPolicyDetails() {
     const policyInfo = await this.secretService.getBtcWalletPolicyDetails();
 
@@ -1142,12 +838,5 @@ export class WalletService implements OnLock, OnUnlock {
       hmac,
       policy,
     };
-  }
-
-  async getWalletType(id: string) {
-    const walletSecrets = await this.secretService.getWalletAccountsSecretsById(
-      id
-    );
-    return walletSecrets?.secretType;
   }
 }
